Refactor TagSystem
The goal here is to make the Tagging system into a gem so it will possibly be useful to others. Main points is to abstract everything having to do with tags and their relationship out into the TagSystem module. This breaks a lot of things so refer to subsequent commits for explanations