Track my noob attempt at Backbone.js
This took me forever to do because I was reading docs for 0.5.3 and following a tutorial that loaded 0.3.3. Just goes to show tutorials should track package versions and also - I should be writing tutorials jackass. This implementation actually is working finally but is in no way proper backbone so lets get to it