Refactor the filterView and updating repocollections again
Such a pain in the ass. I was using too many event bindings to trigger the fetching. Turns out when implementing back button support through the router it all goes to hell. To many or not enough events firing. Here I'mnow using the router methods to exclusively call .fetch() on our repocollection. The FilterView is only responsible for triggering the routing event. The routing event updates the UI and the data.