Add support for showing singular Repo view via Backbone
This is a pretty big refactor that still needs to be cleaned up. I've changed user/repo references to a singular 'owner'. This means the 'owner' object can be either a user or a repo. It has been quite a pain in the ass to get the app to work with both a collection of repos and showing a singular repo while maintaining cleanly bound events. I'm still working on cleaning it up