Save repository data in redis
No more datamapper or sql for the repository class. This decision was made for speed and simplification - two things that sound like I should be shooting for. Basic setting and getting ability and we store the data in a redis hash with full_name as hash key and json data as values