Store User model on redis
Here I've created an Owner model that will persist in redis. While the User model will remain in mysql and will represent a 'user' of the app itself, while an 'owner' is treated as just the data we are displaying. Note these are essentially the same thing but a User will be used exclusively for authentication and app statistics.