Fix bug where we expect 'undefined' to be the same as 'blank'
In the FilterView I'm monitoring whether or not a user is changing the owner scope via the input fields. For example if a user changes 'plusjade' to 'defunkt' the app registers that as a change. However an owner can be 'blank' in that there is no owner scope. FilterView needs to see that 'blank' is equal to 'blank' and not register a change. So if the owner is undefined such as when using Model.clear() we still want 'undefined' to be the same as 'blank'.